# docs/specs/api.openapi.yaml
openapi: 3.1.0
info:
  title: VedioGen Control API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    P0 创作主链路和内部模型管理后台的控制面 API。
    长任务返回 Run ID，通过事件流或查询接口获取状态。
servers:
  - url: /api/v1
...
paths:
...
  /admin/model-providers:
    get:
      summary: List model providers
      tags: [Model Providers]
      operationId: listModelProviders
      security:
        - adminSession: []
      responses:
        "200":
          description: Model providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelProvider"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Create a model provider
      tags: [Model Providers]
      operationId: createModelProvider
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelProviderInput"
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelProvider"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-providers/{providerId}:
    patch:
      summary: Update a model provider
      tags: [Model Providers]
      operationId: updateModelProvider
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelProviderInput"
      responses:
        "200":
          description: Provider updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelProvider"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-credentials:
    get:
      summary: List redacted credential metadata
      tags: [Model Credentials]
      operationId: listModelCredentials
      security:
        - adminSession: []
      responses:
        "200":
          description: Credential metadata only
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelCredential"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Store a new Provider API Key
      tags: [Model Credentials]
      operationId: createModelCredential
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCredentialRequest"
      responses:
        "201":
          description: Credential written; secret is never returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCredential"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-credentials/{credentialId}/probe:
    post:
      summary: Test a stored credential
      tags: [Model Credentials]
      operationId: probeModelCredential
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CredentialId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Credential probe queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncRun"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-credentials/{credentialId}/rotation:
    post:
      summary: Rotate a Provider API Key
      tags: [Model Credentials]
      operationId: rotateModelCredential
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CredentialId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [secret]
              properties:
                secret:
                  type: string
                  minLength: 1
                  format: password
                  writeOnly: true
      responses:
        "202":
          description: New credential stored and rotation probe queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCredential"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-credentials/{credentialId}/revocation:
    post:
      summary: Revoke a Provider API Key
      tags: [Model Credentials]
      operationId: revokeModelCredential
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/CredentialId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Credential enters draining or revoked state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelCredential"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-deployments:
    get:
      summary: List model deployments
      tags: [Model Deployments]
      operationId: listModelDeployments
      security:
        - adminSession: []
      responses:
        "200":
          description: Model deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelDeployment"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Create a model deployment
      tags: [Model Deployments]
      operationId: createModelDeployment
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelDeploymentInput"
      responses:
        "201":
          description: Deployment created in draft state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelDeployment"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-deployments/{deploymentId}/probe:
    post:
      summary: Run capability probes for a model deployment
      tags: [Model Deployments]
      operationId: probeModelDeployment
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/DeploymentId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [fixtureIds]
              properties:
                fixtureIds:
                  type: array
                  minItems: 1
                  uniqueItems: true
                  items:
                    type: string
      responses:
        "202":
          description: Probe queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncRun"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-deployments/{deploymentId}:
    patch:
      summary: Update a model deployment
      tags: [Model Deployments]
      operationId: updateModelDeployment
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/DeploymentId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModelDeploymentInput"
      responses:
        "200":
          description: Deployment updated and returned to draft when capability fields change
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelDeployment"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-routing/drafts:
    get:
      summary: Get the current routing draft
      tags: [Model Routing]
      operationId: getCurrentRoutingDraft
      security:
        - adminSession: []
      responses:
        "200":
          description: Current routing draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingPolicyDraft"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-routing/versions:
    get:
      summary: List immutable routing policy versions
      tags: [Model Routing]
      operationId: listRoutingPolicyVersions
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Routing policy version page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingPolicyVersionPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Create a routing draft
      tags: [Model Routing]
      operationId: createRoutingDraft
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutingPolicyInput"
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingPolicyDraft"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-routing/{routingDraftId}/validation:
    post:
      summary: Validate a routing draft with fixtures
      tags: [Model Routing]
      operationId: validateRoutingDraft
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/RoutingDraftId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Route fixture validation queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncRun"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-routing/{routingDraftId}/publication:
    post:
      summary: Publish an immutable routing version
      tags: [Model Routing]
      operationId: publishRoutingDraft
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/RoutingDraftId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [changeNote]
              properties:
                changeNote:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        "201":
          description: Immutable route version published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingPolicyVersion"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-routing/{routingPolicyVersionId}/rollback:
    post:
      summary: Roll back by publishing a prior route as a new version
      tags: [Model Routing]
      operationId: rollbackRoutingPolicy
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/RoutingPolicyVersionId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [changeNote]
              properties:
                changeNote:
                  type: string
                  minLength: 1
                  maxLength: 500
      responses:
        "201":
          description: New version created from rollback target
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoutingPolicyVersion"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-playground-runs:
    post:
      summary: Run a fixed-fixture model or routing probe
      tags: [Model Playground]
      operationId: createModelPlaygroundRun
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelPlaygroundRunRequest"
      responses:
        "202":
          description: Playground run queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPlaygroundRun"
        default:
          $ref: "#/components/responses/Error"
  /admin/model-playground-runs/{playgroundRunId}:
    get:
      summary: Get a model playground run
      tags: [Model Playground]
      operationId: getModelPlaygroundRun
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/PlaygroundRunId"
      responses:
        "200":
          description: Redacted probe result and actual routing chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelPlaygroundRun"
        default:
          $ref: "#/components/responses/Error"
...
  /admin/model-invocations/{invocationId}:
    get:
      summary: Get one redacted model invocation with its route chain
      tags: [Model Invocations]
      operationId: getModelInvocation
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/InvocationId"
      responses:
        "200":
          description: Redacted model invocation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelInvocationDetail"
        default:
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    userSession:
      type: http
      scheme: bearer
      bearerFormat: session
      description: P0 local mode injects a fixed local user; external deployment requires authentication.
    adminSession:
      type: http
      scheme: bearer
      bearerFormat: session
      description: Requires model-admin; P0 may restrict the admin route to localhost.
  parameters:
...
    ProviderId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z][a-z0-9-]{0,63}$"
    CredentialId:
      name: credentialId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Current row version or ETag.
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    ModelProvider:
      type: object
      additionalProperties: false
      required: [id, displayName, adapterType, baseUrl, region, enabled, status, rowVersion, createdAt, updatedAt]
      properties:
        id:
          type: string
          pattern: "^[a-z][a-z0-9-]{0,63}$"
        displayName:
          type: string
        adapterType:
          type: string
        baseUrl:
          type: string
          format: uri
        region:
          type: string
        enabled:
          type: boolean
        status:
          enum: [draft, active, disabled]
        rowVersion:
          type: integer
          minimum: 1
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
...